feat(plugin): v1.7.8 — Dev Mode, Observability, Architecture (P6)
Plugin:
- Dev Mode: scripts/dev.sh (--plugin-dir), test-hook.mjs (hook isolation)
- Observability: structured JSON logging (logger.mjs), session health report
- Architecture: MCP separated from npm package, per-scope cooldown, cache auto-cleanup
- 9 hooks all instrumented with structured logging
- New skills: /prismer:debug-log, /prismer:plugin-dev
- 81/81 tests passing

Also syncs latest SDK changes (mcp, golang, python, rust, typescript).

Made-with: Cursor

+++ sdk/prismer-cloud/python/prismer/types.py
@@ -117,7 +117,7 @@ class ParseDocument(BaseModel):
     page_count: int = Field(alias="pageCount")
     metadata: Optional[Dict[str, Any]] = None
     images: Optional[List[ParseDocumentImage]] = None
-    estimated_time: Optional[int] = Field(default=None, alias="estimatedTime")
+    estimated_time: Optional[float] = Field(default=None, alias="estimatedTime")
 
     class Config:
         populate_by_name = True
